% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_WHO2016.R
\name{check_data}
\alias{check_data}
\title{Clean and check data collected using WHO 2016 questionnaire}
\usage{
check_data(vadata)
}
\arguments{
\item{vadata}{data frame of the WHO2016 data in the InterVA-5 input format.}
}
\value{
a data frame of symptoms after checking for inconsistencies using the InterVA-5 logic.
}
\description{
Clean and check data collected using WHO 2016 questionnaire
}
\examples{
require(InterVA5)
data(RandomVA5)
cleaned <- check_data(RandomVA5)
cleaned[1:10, 1:10]
}
