% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CODgroup.R
\docType{data}
\name{CODgroup}
\alias{CODgroup}
\title{Mapping between the WHO cause list and broad categories}
\format{
An object of class \code{data.frame} with 61 rows and 2 columns.
}
\usage{
data(CODgroup)
}
\description{
A data frame of two columns that maps WHO cause list into broader categories.
}
\keyword{datasets}
