% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_symptom.R
\name{evaluate_symptom}
\alias{evaluate_symptom}
\title{Create skip pattern list to account for fixed symptom dependence}
\usage{
evaluate_symptom(data, symptom, causes.compute = NULL)
}
\arguments{
\item{data}{data frame of the WHO2016 data in the InterVA-5 input format.}

\item{COD}{data frame with ID and cause of death.}

\item{verbose}{whether to print the symptom tree structures for diagnostics.}

\item{remove.injury}{whether to remove injury related symptoms}
}
\value{
a list of binary data, symptom table, and dependence structure.
}
\description{
Create skip pattern list to account for fixed symptom dependence
}
\examples{
require(InterVA5)
data(RandomVA5)
random_COD <- data.frame(ID = RandomVA5$ID)
set.seed(1)
random_COD$cause = sample(c("Cause 1", "Cause 2"), 
							dim(random_COD)[1], replace=TRUE)
data <- organize_dependence(RandomVA5, COD = random_COD)
metric <- evaluate_symptom(data = data, symptom = "i147o")
metric
}
