% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odkItemResponse.R
\name{findRoot}
\alias{findRoot}
\title{Attempt to find root questions in the ODK Instrument.}
\usage{
findRoot(odk_form, ignore = c("Id10019", "Id10310"), drop_calculate = FALSE)
}
\arguments{
\item{odk_form}{A data frame with the ODK instrument}

\item{ignore}{A vector of strings for item names to ignore when considering
as determined by the relevant field.  Any items that include 'age' or 'is'
are automatically ignored.  There is also an attempt to identify (and ignore)
the item indicating the respondent's consent.}

\item{drop_calculate}{Logical for including calculated items as potential root questions.}
}
\value{
A vector of strings with item names of the root questions.
}
\description{
Makes several assumptions, probably makes a few mistakes.
}
