% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_skip_pattern.R
\name{get_skip_pattern}
\alias{get_skip_pattern}
\title{Create skip pattern list to account for fixed symptom dependence}
\usage{
get_skip_pattern(data, COD, verbose = FALSE, remove.injury = TRUE)
}
\arguments{
\item{data}{data frame of the WHO2016 data in the InterVA-5 input format.}

\item{COD}{data frame with ID and cause of death.}

\item{verbose}{whether to print the symptom tree structures for diagnostics.}

\item{remove.injury}{whether to remove injury related symptoms}
}
\value{
a list of binary data, symptom table, and dependence structure.
}
\description{
Create skip pattern list to account for fixed symptom dependence
}
\examples{
library(openVA)
data(RandomVA5)
random_COD <- data.frame(ID = RandomVA5$ID, cause = "Unknown")
data <- get_skip_pattern(RandomVA5, COD = random_COD)
head(data$symptoms)
head(data$causes)
head(data$ID)
summary(data$dependence)
}
