% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odkItemResponsePlots.R
\name{heatMap}
\alias{heatMap}
\title{Heatmap for selected response items.}
\usage{
heatMap(items, first = 1, last = 30, label_length = 150)
}
\arguments{
\item{items}{A data frame returned from \code{itemMissing}.}

\item{first}{Index for the first item to include in the plot.}

\item{last}{Index for the last item to include in the plot.}

\item{label_length}{Number of characters in label per line.}
}
\description{
\code{heatMap} is intended to be used with the Items data frame produced
by \code{itemMissing}.
}
\examples{
\dontrun{
## Example with 2016 WHO VA instrument version 1.5.1
library(vaItemResponse)
data(who2016_151_data)
data(who2016_151_instrument)
results <- itemMissing(who2016_151_data, who2016_151_instrument)
}

}
