% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapICD.R
\name{mapICD}
\alias{mapICD}
\title{Map ICD-10 codes into the WHO cause list}
\usage{
mapICD(icd10Code)
}
\arguments{
\item{icd10Code}{one ICD-10 codes (instead of a vector)}
}
\description{
Map ICD-10 codes into the WHO cause list
}
\examples{
mapICD("A90")
}
