% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odkItemResponsePlots.R
\name{plotItems}
\alias{plotItems}
\title{Histograms for the item response patterns.}
\usage{
plotItems(items, response)
}
\arguments{
\item{items}{A data frame returned from \code{itemMissing}.}

\item{response}{Name of the column in the data frame to plot.}
}
\description{
\code{plotItems} is intended to be used with the Items data frame produced
by \code{itemMissing}.
}
\examples{
\dontrun{
## Example with 2016 WHO VA instrument version 1.5.1
library(vaItemResponse)
data(who2016_151_data)
data(who2016_151_instrument)
results <- itemMissing(who2016_151_data, who2016_151_instrument, add_percent = TRUE)

}

}
