% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_ranking.R
\name{summary_ranking}
\alias{summary_ranking}
\title{Map ICD-10 codes into the WHO cause list}
\usage{
summary_ranking(metric, data)
}
\arguments{
\item{metric}{a data frame of item significance metric with at least the following three columns: cause, symptom, value. This can be the combined output from evaluate_symptom function.}

\item{data}{output of the organize_dependence function.}
}
\description{
Map ICD-10 codes into the WHO cause list
}
\examples{
\dontrun{
require(InterVA5)
data(RandomVA5)
random_COD <- data.frame(ID = RandomVA5$ID)
set.seed(1)
random_COD$cause = sample(c("Cause 1", "Cause 2"), 
							dim(random_COD)[1], replace=TRUE)
data <- organize_dependence(RandomVA5, COD = random_COD)
metric_all <- NULL
for(s in  data$symptoms[1:10, 1]){
 metric <- evaluate_symptom(data = data, symptom = s)
  metric_all <- rbind(metric_all, metric)
}
summary <- summary_ranking(metric_all, data)
summary$symptom[["i022a"]]
summary$table
summary$percentile
}
}
