% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odkItemResponse.R
\name{translate}
\alias{translate}
\title{Translate ODK relevant field}
\usage{
translate(relevant)
}
\arguments{
\item{relevant}{A string from ODK$relevant that needs to be translated.}
}
\description{
\code{translate} converts conditions in ODK to R.
}
\details{
This is a worker function.
}
\examples{
\dontrun{
## Example with 2016 WHO VA instrument version 1.5.1
library(vaItemResponse)
data(who2016_151_instrument)
relevant <- who2016_151_instrument$relevant[15]
translate(relevant)
}

}
