% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odkData.R
\docType{data}
\name{who2016_151_data}
\alias{who2016_151_data}
\title{Example data generated from the WHO 2016 verbal autopsy instrument (version 1.5.1).}
\format{
An object of class \code{data.frame} with 139 rows and 527 columns.
}
\usage{
data(who2016_151_data)
}
\description{
A data frame with 139 rows (deaths) and 527 columns (items), similar to an Open
Data Kit (ODK) export file.
}
\keyword{datasets}
