% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odkData.R
\docType{data}
\name{who2016_151_instrument}
\alias{who2016_151_instrument}
\title{The WHO 2016 verbal autopsy instrument (version 1.51).}
\format{
An object of class \code{data.frame} with 588 rows and 13 columns.
}
\usage{
data(who2016_151_instrument)
}
\description{
A data frame that contains the "survey" tab in excel workshop for the
WHO 2016 verbal autopsy instrument (version 1.5.1).  For more information
see:
}
\details{
https://www.who.int/healthinfo/statistics/verbalautopsystandards/en/

For the latest version of the instrument, see:

https://github.com/SwissTPH/WHO_VA_2016
}
\keyword{datasets}
